Moved saveContainer to LocalContainerProvider. There's a function getDefaultSaveProvider which should be expanded accordingly in the future.
Contributes to issue CURA-4243.
